requirements.txt
# Base
python-dotenv==1.0.1
pyyaml==6.0.2

# Data Science stack
pandas==1.5.3
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2

# App framework
streamlit==1.38.0

# IA e embeddings
openai==0.28.1

faiss-cpu==1.8.0.post1
langchain==0.2.14
sentence-transformers==3.2.0

# PandasAI (versão estável, sem dependências Rust)
pandasai==2.0.0
polars==0.20.16

# Utilitários
tqdm==4.66.5
